+++ test/clj_kondo/main_test.clj
@@ -857,7 +857,8 @@ foo/foo ;; this does use the private var
   (is (empty?
        (lint! "(defn foo [[x y] {:keys [:z] :or {z (+ x y)}}] z)"
               {:linters {:unresolved-symbol {:level :error}
-                         :unused-binding {:level :warning}}}))))
+                         :unused-binding {:level :warning}
+                         :improper-or-mapping {:level :off}}}))))
 
 (deftest let-test
   (assert-submap